requirements.txt
absl-py==0.15.0
anaconda-client==1.7.2
anaconda-navigator==2.1.0
aniso8601==9.0.1
appdirs==1.4.4
argh==0.26.2
astunparse==1.6.3
atomicwrites==1.4.0
autocrop==1.1.1
Automat==20.2.0
backcall==0.2.0
backports.weakref==1.0.post1
bkcharts==0.2
black==19.10b0
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
clang==5.0
click==8.0.3
clyent==1.2.2
conda==4.11.0
conda-build==3.21.4
conda-verify==3.4.2
constantly==15.1.0
cssselect==1.1.0
cycler==0.10.0
cytoolz==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
et-xmlfile==1.1.0
fastcache==1.1.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gmpy2==2.0.8
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
gunicorn==20.1.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
htmlmin==0.1.12
hyperframe==5.2.0
hyperlink==21.0.0
incremental==21.3.0
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
itemadapter==0.4.0
itemloaders==1.0.4
jedi==0.18.0
jmespath==0.10.0
jsonify==0.5
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-core==4.8.1
kaggle==1.5.12
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
llvmlite==0.36.0
locket==0.2.1
Markdown==3.3.4
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
multipledispatch==0.6.0
mypy-extensions==0.4.3
navigator-updater==0.2.1
nest-asyncio==1.5.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
pandas==1.3.3
parsel==1.6.0
parso==0.8.2
pathspec==0.7.0
patsy==0.5.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pkginfo==1.7.1
plotly==5.5.0
ply==3.11
priority==1.3.0
prompt-toolkit==3.0.21
Protego==0.1.16
protobuf==3.19.0
ptyprocess==0.7.0
pyaaf2==1.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycurl==7.44.1
PyDispatcher==2.0.5
Pygments==2.10.0
pyodbc===4.0.0-unsupported
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pytest==6.2.4
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
QDarkStyle==2.8.1
queuelib==1.6.2
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
Scrapy==2.5.1
service-identity==21.1.0
simplegeneric==0.8.1
sip==4.19.13
six==1.15.0
split-folders==0.4.3
tables==3.6.1
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.4
text-unidecode==1.3
tifffile==2020.10.1
tornado==6.1
traitlets==5.1.0
tweepy==4.4.0
Twisted==21.7.0
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.7
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xlwt==1.3.0
zict==2.0.0
zope.event==4.5.0
